backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
slowapi==0.1.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
cryptography==41.0.7
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# ═════════════════════════════════════════════════════════════
# DATABASE DRIVERS
# ═════════════════════════════════════════════════════════════
# psycopg2-binary = "Traducteur" Python ↔ PostgreSQL
#   Sans ça, Python peut pas parler à la base de données
#   Requis pour: engine = create_engine("postgresql://...")
psycopg2-binary==2.9.9

# ═════════════════════════════════════════════════════════════
# AUTHENTICATION & SECURITY
# ═════════════════════════════════════════════════════════════
jose[cryptography]==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
